Make SymbolHistogram non-generic, fix henikoff weight calculation

+++ rust/weighted_ld/src/main.rs
@@ -136,7 +136,7 @@ fn main() -> Result<(), std::io::Error> {
     );
 
     let sw = Instant::now();
-    let min_acgt = (opt.min_acgt * siteset.n_seqs() as f32).ceil() as u32;
+    let min_acgt = (opt.min_acgt * siteset.n_seqs() as f32).ceil() as usize;
     let min_minor = opt.min_minor;
     let max_minor = opt.max_minor;
     let filtered_siteset =
